requirements.txt
# Versões testadas e compatíveis
weasyprint==60.2
pydyf==0.10.0
cssselect2==0.7.0
tinycss2==1.3.0
cffi==1.16.0
pillow==10.1.0
flask==3.0.3
markdown2==2.4.13
gunicorn==21.2.0

# AI Dependencies
openai==1.3.9
openai-whisper==20231117
torch
torchaudio
ffmpeg-python==0.2.0